refactor(shell): isolate admin playlist route boundary
